rename notifier processor
